# perf: aggressively cache getUsersCompletedGamesAndMax