This PR adds caching to the Awarded data retrieved from the `getUsersCompletedGamesAndMax()` call on user profiles.

The cache is populated on page load, and busted when the user unlocks an achievement or resets progress.

If cached data is present, a lightweight query runs instead and the cached data is merged onto the lightweight fetch results.

### Benchmarks
These benchmarks are taken with the local MySQL cache disabled.

User: televandalist
|         | Before | After |
|---------|--------|-------|
| 1       | 421ms  | 11ms  |
| 2       | 410ms  | 10ms  |
| 3       | 432ms  | 10ms  |
| 4       | 423ms  | 11ms  |
| 5       | 425ms  | 11ms  |
| Average | 422ms  | 11ms  |

Speedup Factor: ~38x

---

User: CobaltZeroni
|         | Before | After |
|---------|--------|-------|
| 1       | 483ms  | 31ms  |
| 2       | 493ms  | 31ms  |
| 3       | 619ms  | 29ms  |
| 4       | 601ms  | 32ms  |
| 5       | 512ms  | 53ms  |
| Average | 542ms  | 35ms  |

Speedup Factor: ~15x